Move stringutils from internal to pkg
