+++ pkg/ignore/ignore.go
@@ -1,7 +1,7 @@
 package ignore
 
 import (
-	"golang.stackrox.io/kube-linter/internal/stringutils"
+	"golang.stackrox.io/kube-linter/pkg/stringutils"
 )
 
 const (
