+++ pkg/stringutils/consume.go
