+++ pkg/stringutils/ternary.go
