+++ pkg/templates/nonexistentserviceaccount/template.go
@@ -4,13 +4,13 @@ import (
 	"fmt"
 
 	"golang.stackrox.io/kube-linter/internal/set"
-	"golang.stackrox.io/kube-linter/internal/stringutils"
 	"golang.stackrox.io/kube-linter/pkg/check"
 	"golang.stackrox.io/kube-linter/pkg/config"
 	"golang.stackrox.io/kube-linter/pkg/diagnostic"
 	"golang.stackrox.io/kube-linter/pkg/extract"
 	"golang.stackrox.io/kube-linter/pkg/lintcontext"
 	"golang.stackrox.io/kube-linter/pkg/objectkinds"
+	"golang.stackrox.io/kube-linter/pkg/stringutils"
 	"golang.stackrox.io/kube-linter/pkg/templates"
 	"golang.stackrox.io/kube-linter/pkg/templates/nonexistentserviceaccount/internal/params"
 	v1 "k8s.io/api/core/v1"
